+++ core/foundation/foundation.hpp
@@ -16,6 +16,7 @@
 #include "result.hpp"
 #include "status.hpp"
 #include "stopwatch.hpp"
+#include "string.hpp"
 #include "time.hpp"
 #include "timestamp.hpp"
 #include "uuid.hpp"